Move common part of TemplateDTO and ProcessDTO to BaseTemplateDTO
